% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS2repo-package.R
\docType{package}
\name{DS2repo-package}
\alias{DS2repo}
\alias{DS2repo-package}
\title{DS2repo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
