% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_by_n.R
\name{multiply_by_n}
\alias{multiply_by_n}
\title{Multiply a vector by a scalar "n"}
\usage{
multiply_by_n(vector, n)
}
\arguments{
\item{vector}{Numeric vector}

\item{n}{Numeric scalar}
}
\value{
The product of the vector multiplied by the scalar N
}
\description{
This function multiplies all the elements of a vector by the scalar N.
Both arguments must comprise numeric elements and must be defined by the
user.
}
\examples{
multiply_by_n(vector = c(1,2,3,4),
              n = 5)

}
